requirements.txt
aiofiles
aiohttp
apscheduler
asyncio
bard 
beautifulsoup4
bs4
cloudscraper
daxxhub
dnspython
ffmpeg-python
Faker
future==1.0.0
git+https://github.com/AsmSafone/SafoneAPI
gitpython
gtts
hachoir
heroku3
httpx==0.27.0
img2pdf
instaloader
lexica-api
motor 
MukeshAPI
numpy<2.0.0
openai==0.28.0
opencv-python
psutil
pickledb
pycountry
pycryptodome
pydub
pyfiglet
pykeyboard
pyrogram==2.0.106
python-dotenv
python-telegram-bot==21.1.1
python-whois
pyshorteners
pyttsx3
pytz
pytube
py-tgcalls==0.9.7
pyyaml==6.0.1
qrcode
requests
SafoneAPI
search_engine_parser
speedtest-cli
SpeechRecognition
spotipy
jiosaavnpy
tgcrypto
unidecode
wget
yt-dlp
youtube-search
youtube-search-python
flask
mutagen
